A proper home for the (updated and corrected) history content (#3)
I revived the history section off the old Roadmap page, and updated some of it to reflect reality (some of it was wrong) based on a review of historical Mozilla documents with the help of Asa Dotzler. It links to source references where available (old Mozilla press releases, commit logs, etc). Also putting it in a more appropriate place (the About page).